fix(perf): guard pathname cookie and cache token list fetches server-side
- Guard pathname cookie in middleware — set only on first page load
  instead of every request, removing the unconditional Set-Cookie header
  that forced cf-cache-status: BYPASS on all HTML responses
- Extract getTokensQuery / getChainTokensQuery from useTokens into
  src/app/lib/tokens/tokenQueries.ts
- Add cachedTokensFetch.ts (unstable_cache, revalidate 300s) to bypass
  the 2 MB Next.js fetch-cache cap on the 7 MB token list; sitemap
  bridge generator and widget-amounts image route now use cached helpers
